% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_model.r
\name{choose_model}
\alias{choose_model}
\title{Function assessing different types of regression models, returing the one
giving the best results according to an information criterion.}
\usage{
choose_model(dataset, dependent, independents)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}
}
\value{
A list with named elements 'model', 'type', 'score'
}
\description{
Function assessing different types of regression models, returing the one
giving the best results according to an information criterion.
}
\examples{
my_model <- choose_model(d, y, x)$model
}
