% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_removal.R
\name{choose_removal}
\alias{choose_removal}
\title{Finds the best knot to remove from the given model}
\usage{
choose_removal(
  dataset,
  dependent,
  independents,
  knots,
  boundary_knots,
  cost_fn = stats::BIC
)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{knots}{The knot placements}

\item{boundary_knots}{The boundary knot placements}

\item{cost_fn}{The function for the selection criterion score (BIC default)}
}
\value{
A named list with 'index' of the chosen knot, 'model', and 'score'
}
\description{
The function searches through the knots of the given model to find the knot
yielding the best resulting score if one knot has to be removed.
The selection criterion function can be seen as a cost function, where
lower scores are better.
}
\examples{
removal_index <- choose_removal(my_data, y, x, knots, boundary_knots)$index
removal_index <- choose_removal(my_data, y, x, knots, boundary_knots,
 cost_fn = function(model) { return(-summary(model)$fstatistic[[1]]) })$index
}
