% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_splines_thorough.R
\name{choose_splines_thorough}
\alias{choose_splines_thorough}
\title{Uses several approaches to find the best of a set of natural cubic splines
regression models with a knot count lower than or equal to a specified
maximum number of knots.}
\usage{
choose_splines_thorough(
  d,
  dependent,
  independents,
  max_nknots = 7,
  icr_fn = stats::BIC,
  cost_fn = stats::AIC
)
}
\arguments{
\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{max_nknots}{The maximum number of knots wanted}

\item{icr_fn}{The information criterion function comparing models with
different knot counts (BIC default)}

\item{cost_fn}{The criterion used to choose which knot to remove, used by
the function 'choose_removal'. Default is AIC.}

\item{dataset}{The data frame}
}
\value{
List of models
}
\description{
The maximum number of knots is given as an input argument.
}
