% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_knots.R
\name{extract_knots}
\alias{extract_knots}
\title{Extracts the distinct knot placements of natural spline regression models}
\usage{
extract_knots(ns_model)
}
\arguments{
\item{ns_model}{The natural spline regression model}
}
\value{
A list with named elements 'knots' and 'Boundary.knots'
}
\description{
Extracts the distinct knot placements of natural spline regression models
}
\examples{
extract_knots(my_model)
}
