% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{f_noise_rnorm}
\alias{f_noise_rnorm}
\title{Function for the noise (variance) to add to the signal that can be passed to
'generate_data' as the 'f_noise' parameter.}
\usage{
f_noise_rnorm(xs)
}
\arguments{
\item{xs}{The x (independent) values for which to generate noise/randomness}
}
\value{
The noise/randomness to add to the y values (dependent variable's
population means)
}
\description{
This function is based on drawing a values from a normal distribution.
}
