% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{f_signal_cos}
\alias{f_signal_cos}
\title{Function for the relationship between independent and dependent variable,
which can be passed to 'generate_data' as the 'f_signal' parameter.}
\usage{
f_signal_cos(xs)
}
\arguments{
\item{xs}{The x values}
}
\value{
The raw signal part of the y values
}
\description{
This signal is based on a cosine curve.
}
