% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{f_signal_michaelis_menten}
\alias{f_signal_michaelis_menten}
\title{Function for the relationship between independent and dependent variable,
which can be passed to 'generate_data' as the 'f_signal' parameter.}
\usage{
f_signal_michaelis_menten(xs)
}
\arguments{
\item{xs}{The x values}
}
\value{
The raw signal part of the y values
}
\description{
This signal is based on the Michaelis-Menten equation with
K_m = 2.0 μM, V_max = 10.0 μmol/sec.
}
