% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{f_signal_x_div_half_plus_x}
\alias{f_signal_x_div_half_plus_x}
\title{Function for the relationship between independent and dependent variable,
which can be passed to 'generate_data' as the 'f_signal' parameter.}
\usage{
f_signal_x_div_half_plus_x(xs)
}
\arguments{
\item{xs}{The x values}
}
\value{
The raw signal part of the y values
}
\description{
This signal is based on the formula f(x) = x / (0.5 + x)
}
