% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{f_x_rnorm_dist}
\alias{f_x_rnorm_dist}
\title{A function that can be passed to 'generate_data' as the 'f_x_dist' parameter.
Normally distributed.}
\usage{
f_x_rnorm_dist(n)
}
\arguments{
\item{n}{The number of units in the sample}
}
\value{
The raw values for x
}
\description{
A function that can be passed to 'generate_data' as the 'f_x_dist' parameter.
Normally distributed.
}
