% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generates synthetic data}
\usage{
generate_data(n, x_accr, y_accr, f_x_dist, f_signal, f_noise)
}
\arguments{
\item{n}{The number of units in the sample}

\item{x_accr}{The accuracy (number of decimals) of independent variable}

\item{y_accr}{The accuracy (number of decimals) of dependent variable}

\item{f_x_dist}{Function for the distribution of the independent variable}

\item{f_signal}{The true function for the population means. Takes the x
values as input.}

\item{f_noise}{The noise function, i.e., variance distribution(s) around the
population means for y, which can be heteroscedastic.}
}
\value{
A dataframe with generated data
}
\description{
The returned data frame has the following columns:
ID               # IDs
Independent      # The measured X values (rounded according to accuracy)
Dependent        # The measured Y values (rounded according to accuracy)
IndependentRaw   # The raw X values (not rounded)
DependentRaw     # The raw Y values (not rounded)
SignalRaw        # The raw signal part of the population mean curve for Y
Noise            # The raw noise (variance) for the DependentRaw values
SignalMeasured   # The signal rounded according to accuracy
}
