% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_splines_thorough.R
\name{is_empty_interval}
\alias{is_empty_interval}
\title{Check if the subinterval <a, b] for a variable is empty}
\usage{
is_empty_interval(d, v, a, b)
}
\arguments{
\item{d}{The dataset}

\item{v}{The variable}

\item{a}{The lower, exclusive limit}

\item{b}{The upper, inclusive limit
return TRUE if empty, FALSE if non-empy}
}
\description{
Check if the subinterval <a, b] for a variable is empty
}
