% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_splines.R
\name{model_by_count}
\alias{model_by_count}
\title{Utility function for natural splines that wraps the 'glm' and 'ns' functions}
\usage{
model_by_count(dataset, dependent, independents, nknots)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{nknots}{The requested number of knots}
}
\value{
The regression model
}
\description{
Creates a natural splines regression model given the wanted number of knots
}
\examples{
my_model <- create_model(my_data, y, x, 7)
}
