% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted_cubic_splines.R
\name{model_by_knots}
\alias{model_by_knots}
\title{Utility function for restricted cubic splines that wraps the 'glm' and 'ns'
functions}
\usage{
model_by_knots(dataset, dependent, independents, knots, boundary_knots)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{knots}{The inner knot placements}

\item{boundary_knots}{The boundary knot placements}
}
\value{
The regression model
}
\description{
Creates a restricted cubic spline regression model from the given placements
of the inner knots and boundary knots
}
\examples{
my_model <- model_by_knots(my_data, y, x, c(0.1, 0.2), c(0.0, 0.3))
}
