% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plots the (glm) model with data, curve, confidence bands, etc.}
\usage{
plot_model(dataset, dependent, independent, model)
}
\arguments{
\item{dataset}{The data frame}

\item{mod}{The model}
}
\description{
Plots the (glm) model with data, curve, confidence bands, etc.
}
\examples{
plot_model(d, y, x, my_mod)
}
