% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rstandard.R
\name{plot_rstandard}
\alias{plot_rstandard}
\title{Plots the standardized residuals for the model}
\usage{
plot_rstandard(dataset, model)
}
\arguments{
\item{dataset}{The data frame}

\item{mod}{The model}
}
\description{
Plots the standardized residuals for the model
}
\examples{
plot_rstandard(d, my_mod)
}
