% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_knots.R
\name{print_knots}
\alias{print_knots}
\title{Print knot placements}
\usage{
print_knots(knot_placements)
}
\arguments{
\item{knot_placements}{A list with two named lists, 'knots' and
'Boundary.knots'}
}
\description{
Print knot placements
}
\examples{
print_knots(knot_placements)
}
