% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_splines_thorough.R
\name{remove_knots}
\alias{remove_knots}
\title{Find best model with less than 'max_nknots' given the dataset, knots,
boundary knots, information criterion, and a cost function criterion.}
\usage{
remove_knots(
  d,
  dependent,
  independents,
  knots,
  b_knots,
  max_nknots = 7,
  icr_fn = stats::BIC,
  cost_fn = stats::AIC
)
}
\arguments{
\item{d}{The dataset}

\item{dependent}{The dependent variable in the forumla}

\item{independents}{The independent variables in the formula}

\item{knots}{The knot placements}

\item{b_knots}{The lower and upper boundary knots}

\item{max_nknots}{The maximum number of knots wanted in the final model}

\item{icr_fn}{The information criterion for choosing best model}

\item{cost_fn}{A cost function used to evaluate which knot to remove.}
}
\value{
The best scoring model with maximum 'max_nknots' knots
}
\description{
Find best model with less than 'max_nknots' given the dataset, knots,
boundary knots, information criterion, and a cost function criterion.
}
