% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_knotcount.R
\name{suggest_knotcount}
\alias{suggest_knotcount}
\title{Finds the number of natural spline knots that gives the lowest score for an
information criterion and a data set.}
\usage{
suggest_knotcount(
  dataset,
  dependent,
  independents,
  max_nknots = -1,
  info_crit = stats::AIC
)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variable(s) in the formula}

\item{info_crit}{The information criterion function. Defaults to AIC}

\item{maximum_knots}{The highest knot count to assess. Defaults to 300}
}
\value{
A list with named elements 'nknots' and 'score'
}
\description{
Finds the number of natural spline knots that gives the lowest score for an
information criterion and a data set.
}
\examples{
suggest_knotcount(d, nwsize, age_dec)
}
