% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_model.R
\name{suggest_model}
\alias{suggest_model}
\title{Suggests a model with knot placements that usually are non-uniform with
respect to quantiles and widths}
\usage{
suggest_model(
  dataset,
  dependent,
  independents,
  target_nknots,
  initial_nknots = -1,
  cost_fn = stats::AIC
)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{target_nknots}{The target and maximum number of knots for the model}

\item{initial_nknots}{The number of knots initially, defaults to the
result from the suggest_knotcount-function}

\item{cost_fn}{The function for the selection criterion score (AIC default)}
}
\value{
The suggested regression model
}
\description{
The target number of knots for the model is given as a parameter. The
algorithm starts with a regression model with a high number of knots and
systematically removes knots until the target number of knots is reached.
The initial number of knots can be given as a parameter, and defaults to
half of the number of rows in the dataset, using integer division by two.
}
\examples{
my_model <- suggest_model(d, y, x, 7)
my_model <- suggest_model(d, y, x, 7, 300, BIC)
}
