% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_splines.R
\name{suggest_splines}
\alias{suggest_splines}
\title{Suggests a restricted cubic splines regression model with inner knot
placements that can be non-uniform with respect to quantiles and widths}
\usage{
suggest_splines(
  dataset,
  dependent,
  independents,
  target_nknots,
  ...,
  initial_nknots = -1,
  cost_fn = stats::BIC,
  all_knots = FALSE,
  boundary_knots = NA
)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{target_nknots}{The target and maximum number of inner knots for the
model}

\item{initial_nknots}{The number of inner knots initially, defaults to the
result from the function 'suggest_knotcount'}

\item{cost_fn}{The function for the selection criterion score (BIC default)
used to compare which inner knot should be removed, passed to choose_removal}

\item{all_knots}{If TRUE, then knots for all intermediate models will be
included in return value. Default is FALSE.}

\item{boundary_knots}{The boundary knot placements or NA if not specified}
}
\value{
Returns the suggested natural splines model, or if the 'all_knots'
argument was TRUE, then a list with named elements 'model', 'all_knots', and
'Boundary.knots' is returned.
}
\description{
The target number of inner knots for the model is given as an input argument.
The algorithm starts with a model with a high number of knots and
systematically removes inner knots until the target number of inner knots is
reached.
The initial number of inner knots, before starting to remove knots, can be
given as an argument as well, but  defaults to the suggested number
obtained from the function 'suggest_knotcount'.
}
\examples{
my_model <- suggest_splines(d, y, x, 4)
my_model <- suggest_splines(d, y, x, 4, initial_nknots = 100, cost_fn = BIC)
}
