% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_model.R
\name{choose_model}
\alias{choose_model}
\title{Function assessing different types of regression models, returning the one
giving the best results according to an information criterion}
\usage{
choose_model(
  dataset,
  dependent,
  independents,
  icr_fn = stats::BIC,
  fp_alpha = NA,
  max_nsknots = 7,
  max_fp_df = 4,
  verbose = TRUE
)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{icr_fn}{The information criterion function (default AIC)}

\item{fp_alpha}{The relax factor for multivariate fractional polynomials}

\item{max_nsknots}{The max number of knots for natural splines (default 7)}

\item{max_fp_df}{The max degrees of freedom for fractional polynomials
(default 7)}

\item{verbose}{Verbose output, default TRUE}
}
\value{
A list with named elements, such as 'model', 'type', 'score'. The
function returns a list with named elements and sublists, see examples for
full overview of the returned values.
}
\description{
Function assessing different types of regression models, returning the one
giving the best results according to an information criterion
}
\examples{
my_model <- choose_model(d, y, x)$model
result <- choose_model(d, y, x, icr_fn = AIC, verbose = FALSE)

ret <- choose_model(d, y, x)

ret$labels[[ret$type]] # Human readable name of chosen model type
ret[[ret$type]]        # Gives more values for the chosen model if available

ret$model      # The chosen model
ret$score      # The chosen model's score
ret$type       # The type of model chosen as a string
ret$score_name # The type of score used as a string
ret$score_fn   # The function used for scores

ret$labels     # Description string for the types of models

ret$mfp        # Multivariate fractional polynomial
ret$mfp$model  # The model
ret$mfp$score  # The score

ret$ns         # Natural splines from quantiles
ret$ns$model   # The model
ret$ns$score   # The score
ret$ns$knot_cnt_arg      # The number of knots (df - 1) as input argument
ret$ns$knot_cnt_distinct # The number of distinct placements in the result
ret$ns$knot_placements   # Knots and boundary knots
ret$ns$knot_placements$knots           # The knot placements as a list
ret$ns$knot_placements$Boundary.knots  # The boundary knots as a list

ret$ns_nu         # Natural splines non-uniform placements
ret$ns_nu$model   # The model
ret$ns_nu$score   # The score
ret$ns_nu$knot_cnt_distinct # The number of distinct placements in the result
ret$ns_nu$knot_placements   # Knots and boundary knots
ret$ns_nu$knot_placements$knots          # The knot placements as a list
ret$ns_nu$knot_placements$Boundary.knots # The boundary knots as a list
}
