% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_model.r
\name{choose_model}
\alias{choose_model}
\title{Function assessing different types of regression models, returing the one
giving the best results according to an information criterion}
\usage{
choose_model(
  dataset,
  dependent,
  independents,
  icr_fn = stats::AIC,
  fp_alpha = NA,
  max_nsknots = 7,
  max_fp_df = 4,
  verbose = TRUE
)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{icr_fn}{The information criterion function (default AIC)}

\item{fp_alpha}{The relax factor for multivariate fractional polynomials}

\item{max_nsknots}{The max number of knots for natural splines (default 7)}

\item{max_fp_df}{The max degrees of freedom for fractional polynomials
(default 7)}

\item{verbose}{Verbose output, default TRUE}
}
\value{
A list with named elements 'model', 'type', 'score'
}
\description{
Function assessing different types of regression models, returing the one
giving the best results according to an information criterion
}
\examples{
my_model <- choose_model(d, y, x)$model
result <- choose_model(d, y, x, icr_fn = BIC, verbose = TRUE)
}
