% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_model.R
\name{choose_model}
\alias{choose_model}
\title{Chooses the best of found models within a range of knot counts.}
\usage{
choose_model(
  dataset,
  dependent,
  independents,
  max_nknots = 10,
  icr_fn = stats::BIC,
  cost_fn = stats::AIC,
  initial_nknots = -1,
  diff_better = 0
)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{max_nknots}{The maximum wanted number of knots}

\item{icr_fn}{The information criterion function (BIC default)}

\item{cost_fn}{For comparing models with equal knot counts (default AIC)}

\item{initial_nknots}{The initial high number of knots for the algorithm
(default is nrow(dataset) \%/\% 2)}

\item{diff_better}{How much lower must the score be for a higher knot count
model to be considered a better model than a lower knot model?}
}
\value{
The suggested chosen 'model', 'score', and 'nknots'
}
\description{
The target number of knots for the model is given as a parameter. The
algorithm starts with a regression model with a high number of knots and
systematically removes knots until the target number of knots is reached.
The initial number of knots can be given as a parameter, and defaults to
half of the number of rows in the dataset, using integer division by two.
}
\examples{
my_model <- suggest_model(d, y, x, 7)
my_model <- suggest_model(d, y, x, 7, 300, BIC)
}
