% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_splines.R
\name{choose_splines}
\alias{choose_splines}
\title{Chooses the best from a set of natural splines regression models with
a knot count lower than or equal to a specified maximum number of konts}
\usage{
choose_splines(
  dataset,
  dependent,
  independents,
  max_nknots = 10,
  icr_fn = stats::BIC,
  cost_fn = stats::AIC,
  initial_nknots = -1,
  diff_better = 0
)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{max_nknots}{The maximum number of knots wanted}

\item{icr_fn}{The information criterion function comparing models with
different knot counts (BIC default)}

\item{cost_fn}{For comparing different models with equal knot counts
(default AIC)}

\item{initial_nknots}{The initial high number of knots for the algorithm
(default is the value from the 'suggest_knotcount'-function)}

\item{diff_better}{How much lower must the score be to make a higher knot
model be deemed a better model than an alternative lower knot model?}
}
\value{
The chosen 'model', 'score', and 'nknots'
}
\description{
The maximum number of knots is given as an input argument.
}
\examples{
my_model <- choose_splines(d, y, x, 7)
my_model <- choose_splines(d, y, x, 7, AIC)
}
