% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_splines_thorough.R
\name{choose_splines_thorough}
\alias{choose_splines_thorough}
\title{Uses several approaches to find the best of a set of natural cubic splines
regression models with a knot count lower than or equal to a specified
maximum number of knots.}
\usage{
choose_splines_thorough(
  d,
  dependent,
  independents,
  max_nknots = 7,
  icr_fn = stats::BIC,
  cost_fn = stats::AIC,
  verbose = TRUE
)
}
\arguments{
\item{d}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{max_nknots}{The maximum number of knots wanted}

\item{icr_fn}{The information criterion function comparing models with
different knot counts (BIC default)}

\item{cost_fn}{The criterion used to choose which knot to remove, used by
the function 'choose_removal'. Default is AIC.}

\item{verbose}{Print output about progress and results (Default TRUE)}
}
\value{
The resulting 'model', 'score', etc. of the best scoring model, and
lists with information about the other assessed models. See examples.
}
\description{
The maximum number of knots is given as an input argument.
}
\examples{
my_model <- choose_splines_thorough(d, y, x)$model
result <- choose_splines_thorough(d, y, x, icr_fn = AIC, verbose = FALSE)

ret <- choose_splines_thorough(d, y, x)

ret$labels[[ret$type]] # Human readable name of chosen model type
ret[[ret$type]]        # Gives more values for the chosen model if available

ret$model      # The chosen model
ret$score      # The chosen model's score
ret$type       # The type of model chosen as a string
ret$score_name # The type of score used as a string
ret$score_fn   # The function used for scores

ret$labels     # Description string for the types of models

ret$quantile        # Model obtained from a start model based on quantiles
ret$quantile$model  # The model
ret$quantile$score  # The score
ret$quantile$nknots     # The number of knots
ret$quantile$knots      # Knots and boundary knots locations
ret$quantile$knots$knots          # Knot locations in a list
ret$quantile$knots$Boundary.knots  # Boundary knot locations in a list

res$uniform
res$distinct
}
