% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_splines_thorough.R
\name{choose_splines_thorough}
\alias{choose_splines_thorough}
\title{Uses several approaches to find the best of a set of natural cubic splines
regression models with a knot count lower than or equal to a specified
maximum number of knots.}
\usage{
choose_splines_thorough(
  d,
  dependent,
  independents,
  max_nknots = 7,
  icr_fn = stats::BIC,
  cost_fn = stats::AIC,
  verbose = TRUE
)
}
\arguments{
\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{max_nknots}{The maximum number of knots wanted}

\item{icr_fn}{The information criterion function comparing models with
different knot counts (BIC default)}

\item{cost_fn}{The criterion used to choose which knot to remove, used by
the function 'choose_removal'. Default is AIC.}

\item{verbose}{Print output about progress and results (Default TRUE)}

\item{dataset}{The data frame}
}
\value{
List of models
}
\description{
The maximum number of knots is given as an input argument.
}
\examples{
my_model <- choose_splines_thorough(d, y, x)$model
result <- choose_splines_thorough(d, y, x, icr_fn = AIC, verbose = FALSE)

ret <- choose_splines_thorough(d, y, x)

ret$labels[[ret$type]] # Human readable name of chosen model type
ret[[ret$type]]        # Gives more values for the chosen model if available

ret$model      # The chosen model
ret$score      # The chosen model's score
ret$type       # The type of model chosen as a string
ret$score_name # The type of score used as a string
ret$score_fn   # The function used for scores

ret$labels     # Description string for the types of models

ret$quantile   # Multivariate fractional polynomial
ret$quantile$model  # The model
ret$quantile$score  # The score

ret$ns         # Natural splines from quantiles
ret$ns$model   # The model
ret$ns$score   # The score
ret$ns$knot_cnt_arg      # The number of knots (df - 1) as input argument
ret$ns$knot_cnt_distinct # The number of distinct placements in the result
ret$ns$knot_placements   # Knots and boundary knots
ret$ns$knot_placements$knots           # The knot placements as a list
ret$ns$knot_placements$Boundary.knots  # The boundary knots as a list

ret$ns_nu         # Natural splines non-uniform placements
ret$ns_nu$model   # The model
ret$ns_nu$score   # The score
ret$ns_nu$knot_cnt_distinct # The number of distinct placements in the result
ret$ns_nu$knot_placements   # Knots and boundary knots
ret$ns_nu$knot_placements$knots          # The knot placements as a list
ret$ns_nu$knot_placements$Boundary.knots # The boundary knots as a list
}
