% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_knots.R
\name{print_knots}
\alias{print_knots}
\title{Function assessing different types of regression models, returning the one
giving the best results according to an information criterion}
\usage{
print_knots(knot_placements)
}
\arguments{
\item{knot_placements}{A list with two named lists, 'knots' and
'Boundary.knots'}
}
\description{
Function assessing different types of regression models, returning the one
giving the best results according to an information criterion
}
\examples{
print_knots(knot_placements)
}
