% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_knotcount.R
\name{suggest_knotcount}
\alias{suggest_knotcount}
\title{Finds the number of natural spline knots that gives the lowest score for an
information criterion and a data set.}
\usage{
suggest_knotcount(
  dataset,
  dependent,
  independents,
  max_nknots = -1,
  icr_fn = stats::BIC,
  all_scores = FALSE
)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variable(s) in the formula}

\item{icr_fn}{The information criterion function. Defaults to BIC}

\item{all_scores}{If TRUE, all scores are returned in a list 'all_scores'}

\item{maximum_knots}{The highest knot count to assess. Defaults the lower of
40 and n / 3}
}
\value{
A list with named elements 'nknots', 'score', and 'all_scores'
}
\description{
Finds the number of natural spline knots that gives the lowest score for an
information criterion and a data set.
}
\examples{
suggest_knotcount(d, nwsize, age_dec)
}
