% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_splines.R
\name{suggest_splines}
\alias{suggest_splines}
\title{Suggests a natural splines regression model with knot placements that
can be non-uniform with respect to quantiles and widths}
\usage{
suggest_splines(
  dataset,
  dependent,
  independents,
  target_nknots,
  initial_nknots = -1,
  cost_fn = stats::AIC
)
}
\arguments{
\item{dataset}{The data frame}

\item{dependent}{The dependent variable in the formula}

\item{independents}{The independent variables in the formula}

\item{target_nknots}{The target and maximum number of knots for the model}

\item{initial_nknots}{The number of knots initially, defaults to the
result from the function 'suggest_knotcount'}

\item{cost_fn}{The function for the selection criterion score (AIC default)}
}
\value{
The suggested natural splines model
}
\description{
The target number of knots for the model is given as an input argument. The
algorithm starts with a model with a high number of knots and systematically
removes knots until the target number of knots is reached.
The initial number of knots, before starting to remove knots, can be given
as an argument as well, but  defaults to the suggested number of knots
obtained from the function 'suggest_knotcount'.
}
\examples{
my_model <- suggest_splines(d, y, x, 7)
my_model <- suggest_splines(d, y, x, 7, 300, BIC)
}
